Move the default autoloaders inflector to railties
